+++ changelog.md
@@ -2,6 +2,8 @@ Codion Change Log
 ==================
 
 ## 0.17.35-SNAPSHOT
+### is.codion.framework.model
+- EntityEditModel.Insert, Update and Delete split into separate steps, related refactoring.
 ### is.codion.swing.framework.model
 - EntityEditModel.setDefaults() renamed defaults().
 - EntityEditModel.setDefault() renamed defaultValue(), now Value based.